always use extensionsController/platformContext in browser extension
This shrinks the delta between code paths in the browser extension for "Use extensions" and non-"Use extensions". It makes it so that the platformContext and extensionsController are always instantiated.

This change is tested to be backwards-compatible with Sourcegraph 2.10. In older Sourcegraph versions with extensions not enabled by default (2.10-2.11 inclusive), the user settings will not contain any extensions, and so the extensions controller will effectively be a noop. It is useful to still call it (even if it's a noop) to shrink the delta between the 2 code paths.